:rotating_light: fix typecript errors

+++ src/pages/api/caught/release.ts
@@ -4,8 +4,12 @@ import { prisma } from '~/lib/prisma';
 
 const handleDelete = async (res: NextApiResponse, req: NextApiRequest) => {
   const { id } = req.query;
+  const singleId = Array.isArray(id) ? id[0] : id;
+  if (typeof singleId !== `string`) {
+    return res.status(400).json({ error: `Invalid ID` });
+  }
   await prisma.caught.delete({
-    where: { id },
+    where: { id: singleId },
   });
   return res.status(200).json({ success: true, message: `Pokémon released` });
 };
